chore: refactor services methods
