+++ src/main/java/com/codechallenge/trackfinder/service/SpotifyApiClientService.java
@@ -1,5 +1,6 @@
 package com.codechallenge.trackfinder.service;
 
+import com.codechallenge.trackfinder.dto.ImageCover;
 import com.codechallenge.trackfinder.dto.spotify.GetAlbumResponse;
 import com.codechallenge.trackfinder.dto.spotify.SearchTrackResponse;
 
@@ -7,4 +8,5 @@ public interface SpotifyApiClientService {
     String getToken();
     SearchTrackResponse searchTrack(String isrc);
     GetAlbumResponse getAlbum(String id);
+    ImageCover getCoverImageFromUrl(String coverUrl, String isrc);
 }
