+++ src/main/java/com/codechallenge/trackfinder/service/impl/SpotifyApiClientServiceImpl.java
@@ -1,6 +1,7 @@
 package com.codechallenge.trackfinder.service.impl;
 
 import com.codechallenge.trackfinder.config.SpotifyApiProperties;
+import com.codechallenge.trackfinder.dto.ImageCover;
 import com.codechallenge.trackfinder.dto.spotify.GetAlbumResponse;
 import com.codechallenge.trackfinder.dto.spotify.SearchTrackResponse;
 import com.codechallenge.trackfinder.dto.spotify.TokenResponse;
@@ -11,6 +12,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.BodyInserters;
 
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.Objects;
 
 @Service
@@ -26,7 +29,59 @@ public class SpotifyApiClientServiceImpl implements SpotifyApiClientService {
         this.webClient = webClient;
     }
 
-    @Override
+    private String getContentType(String coverUrl) {
+        try {
+            URL url = new URL(coverUrl);
+            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+            connection.setRequestMethod("HEAD");
+            connection.connect();
+
+            int responseCode = connection.getResponseCode();
+            if (responseCode != HttpStatus.OK.value()) {
+                throw new IllegalArgumentException("The URL cannot be accessed" + responseCode);
+            }
+
+            String contentType = connection.getContentType();
+            if (contentType == null || !contentType.startsWith("image/")) {
+                throw new IllegalArgumentException("The image could not be downloaded");
+            }
+            return contentType;
+
+        } catch (IllegalArgumentException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new RuntimeException("Unexpected Error downloading the image", e);
+        }
+    }
+
+    public ImageCover getCoverImageFromUrl(String coverUrl, String isrc) {
+        try {
+            String contentType = getContentType(coverUrl);
+
+            byte[] imageBytes = webClient.get()
+                    .uri(coverUrl)
+                    .retrieve()
+                    .bodyToMono(byte[].class)
+                    .block();
+
+            if (imageBytes == null || imageBytes.length == 0) {
+                throw new IllegalArgumentException("The image could not be downloaded from the URL.");
+            }
+
+            return ImageCover.builder()
+                    .coverUrl(coverUrl)
+                    .fileNameCover(isrc)
+                    .contentTypeCover(contentType)
+                    .imageCover(imageBytes)
+                    .build();
+
+        } catch (IllegalArgumentException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new RuntimeException("Unexpected Error downloading the image", e);
+        }
+    }
+
     public String getToken() {
         try {
             TokenResponse response = webClient
@@ -47,7 +102,6 @@ public class SpotifyApiClientServiceImpl implements SpotifyApiClientService {
         }
     }
 
-    @Override
     public SearchTrackResponse searchTrack(String isrc) {
         String token = getToken();
 
@@ -65,7 +119,6 @@ public class SpotifyApiClientServiceImpl implements SpotifyApiClientService {
         }
     }
 
-    @Override
     public GetAlbumResponse getAlbum(String id) {
         String token = getToken();
